requirements.txt
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==22.1.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
blis==0.7.9
Bottleneck==1.3.5
brotlipy==0.7.0
catalogue==2.0.8
category-encoders==2.5.1.post0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
cli-exit-tools==1.2.3.2
click==8.1.3
confection==0.0.3
cramjam==2.6.2
cryptography==38.0.1
cycler==0.11.0
cymem==2.0.7
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastai==2.7.10
fastcore==1.5.27
fastdownload==0.0.7
fastjsonschema==2.16.2
fastparquet==2022.11.0
fastprogress==1.0.3
flit_core==3.6.0
fonttools==4.25.0
fsspec==2022.11.0
idna==3.4
ipykernel==6.15.2
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.1
json5==0.9.6
jsonschema==4.16.0
jupyter_client==7.4.7
jupyter_core==4.11.2
jupyter-server==1.18.1
jupyterlab==3.5.0
jupyterlab-pygments==0.1.2
jupyterlab_server==2.16.3
jupyterlab-widgets==1.0.0
kiwisolver==1.4.2
langcodes==3.3.0
lib-detect-testenv==2.0.2.2
llvmlite==0.39.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
#mistune==0.8.4
#mkl-fft==1.3.1
#mkl-random==1.2.2
#mkl-service==2.4.0
#munkres==1.1.4
#murmurhash==1.0.9
#nb-conda-kernels==2.3.1
#nbclassic==0.4.8
#nbclient==0.5.13
#nbconvert==6.5.4
#nbformat==5.5.0
#nest-asyncio==1.5.5
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numexpr==2.8.4
numpy==1.22.3
opencv-contrib-python==4.6.0.66
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pathlib3x==2.0.2.1
pathy==0.10.0
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
ply==3.11
preshed==3.0.8
prometheus-client==0.14.1
prompt-toolkit==3.0.20
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pycparser==2.21
pydantic==1.10.2
Pygments==2.11.2
pyod==1.0.6
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
setuptools==65.5.0
sip==6.6.2
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.3.2.post1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
stack-data==0.2.0
statsmodels==0.13.5
terminado==0.13.1
thinc==8.1.5
threadpoolctl==2.2.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.1.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.12
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.37.1
widgetsnbextension==3.5.2
